Move all Free-related natural transformations to free package
